+++ src/GridCal/Gui/Main/ConsoleLog.py
@@ -3,7 +3,7 @@
 # Form implementation generated from reading ui file 'ConsoleLog.ui',
 # licensing of 'ConsoleLog.ui' applies.
 #
-# Created: Sat Jan 18 17:08:42 2020
+# Created: Thu Feb 20 21:50:31 2020
 #      by: pyside2-uic  running on PySide2 5.13.2
 #
 # WARNING! All changes made in this file will be lost!